extracted home page to add-wallet page
